tweaks to captions modules
