+++ lib/app/templates/transcription_element_index.html.ejs
@@ -7,6 +7,7 @@
         <button type="button" class="btn btn-lg btn-link"  disabled>
           <span id="processingExample" class="glyphicon glyphicon-refresh glyphicon-refresh-animate" aria-hidden="true"></span>&nbsp;&nbsp;
         </button>
+        <p>STT Engine: <kbd><%= sttEngine %></kbd></p>
         <p><%= description %></p>
       </div>
       <div class="col-xs-3 col-sm-2 cold-md-2 col-lg-1">
@@ -19,6 +20,7 @@
 
       <div class="col-xs-9 col-sm-10 cold-md-10 col-lg-11" id="transcriptionCard">
         <button type="button" class="btn btn-lg btn-link showBtn controls" ><%= title %></button>
+        <p><kbd><%= sttEngine %></kbd></p>
         <p><%= description %></p>
       </div>
       <div class="col-xs-3 col-sm-2 cold-md-2 col-lg-1">
@@ -36,6 +38,7 @@
      <!--    <button type="button" class="btn btn-lg btn-link"  disabled>
           <span id="processingExample" class="glyphicon glyphicon-refresh glyphicon-refresh-animate" aria-hidden="true"></span>&nbsp;&nbsp;
         </button> -->
+        <p><kbd><%= sttEngine %></kbd></p>
         <p><%= description %></p>
 
         <% if(error){ %>
