+++ lib/app/templates/transcription_form_template.html.ejs
@@ -189,10 +189,10 @@
      <div role="tabpanel" class="tab-pane" id="captions">
       <br>
       <label for="languageModel">Captions  <code>.srt</code> file:</label><br>
-      <i> coming soon <code>.srt</code></i>
+      
       <!-- srt -->
       <div class="form-group">
-        <label for="fileUpload">Choose an audio or video file to transcribe</label>
+        <label for="fileUpload">Choose a caption file to <code>.srt</code> to associate with your audio or video file</label>
        <% if(window.process !== undefined ){%>
          <button id="btnElectronInputCaptionFile" class="btn btn-default btn-xs"  class="very-sweet-looking">Choose Caption File</button>
          <input id="inputElectronInputCaptionFile" name="file" type="file" disabled="disabled" style=" visibility: hidden;"/>
